Escape comment text for anti-xss goodness.
This makes sure that comments get properly escaped, to eliminate xss
vulnerabilities.

The only thing that I don't like about this patch is that usually, I
like to store unescaped info in the database, and escape it in the view.
There seem to be no view tests or integration tests going on, though,
and so I've just escaped it before storing it in the database. If this
should be done at the view level, something like cucumber or rspec
integration tests should be set up.