fixed #328 Profile image points to dead url on deletion
The issue here was that the ensure_user_picture method compared a
file path against a URL. The test does not reflect this.

Now filenames are compared assuming that profile pictures are
always local.

This solution might cause a performance issue, since the comparison
is not done in the DB anymore.