Remove this, since selected aspects are now handled in the session hash
Not only this code is unnecessary with the new way selected
aspects are stored, also on some edge cases could confuse the user
for example when having two tabs opened and hitting refresh in one
of them. Expecting selected aspects, but the "a_ids[]"-like url could
instead set the new group of selected aspects.