Fix federation with GnuTLS by passing the ca_file to Typhoeus
How did this ever work? Anyway Diaspora now works with libcurl4-gnutls-dev
which already supports SNI in most distributions and is also the default
for many distros. Everybody should switch to it. Do so by installing it and
then do a gem uninstall typhoeus followed by a bundle to compile it with
GnuTLS