Test for mobile UI single post view
Verify that the correct layout is rendered (rather than no layout
at all) in the mobile UI single post view.

This is a test I forgot to write at 5am when I was so excited about
having the mobile UI working on my Blackberry again (bad me!), so
here it is before it slips my mind. :grin:

This tests the change made to posts_controller in 6abe718227bcc858