bootstrap-sass fixes
* Update interim-bootstrap to use the partials from bootstrap-sass,
since some of the names have been changed (and we're using the
Bootstrap 2 javascript now). This should make it easy to keep the
file up to date with the version of Bootstrap we're using.
* Use interim-bootstrap instead of bootstrap-complete in default.css to
avoid the little problem I ran into with Bootstrap + Blueprint grids
not getting along. :)
* Revert the CSS class change (from .span-24.last to .row) in
layouts/application since interim-bootstrap works around that problem.
* Get rid of .icon-black since it's not needed.
* Remove the 60px padding on the page body since it isn't needed and
causes problems with the single post view.