Actually only load rails admin in production mode
This is what I think 28f8db51e79 was supposed to do.

With the initializer in an "unless Rails.env.production?" block
the entire initializer was being ignored in production mode and
loaded in all other environments, but Rails Admin was still
loading regardless of that. :grin:

This only loads rails admin in production and excludes it from
all other environments. Also keeps non-admins out of /admin_panel
(without the initializer it was free-for-all) and temp fixes #3210
until a new gem is released.