AppsController no longer exists
Remove the AppsController check from the next_page_path helper, since
AppsController no longer exists.

fixes #3280

I'd like to write a helper spec for this, but it'll probably require
some stubbing/mocking and I'm a little too tired to deal with that at
the moment.. so for now let's just fix the problem. :)