Remove facebook opengraph indications from the metas
Checking the validity of the main diaspora* page with the W3C validator, it pointed me that two elements were not correctly closed. The element was present twice and is
<code>
<meta content="Diaspora*" property="og:site_name"></meta>
</code>
I duckduckgoed it and found that this element is useful to indicate to Facebook the site name when like buttons are present.

We do not have like buttons, so I propose to remove it. I just want to know if it is linked with the Facebook cross post feature?