Adding hover icoons for notifications & conversations
Added white versions of the icons on hover to match the hover of other
header items.
Changing the notifications hover icon to the regular notifications icon when active so
it is visible with the white background.

Fixing the hard coded background-position value for notifications_grey

Can not depend on generated sprite background position as it will change
with changes in icons

Moving the $icons variable to a broader scope

Since the variable is used in the header, the sprites now are imported
before the header

Using the default sprite map instead of the explicitly created sprite map