Started converting the help section into a backbone view. Things missing are: some urls which were calculated and inserted into the yml, all the pre-existing erb templates are still there, it's still at /faq/faq, nothing has been tested, some of the questions don't collapse and expand.
changed the locales to have the hard coded urls in them because they can no longer be calculated server side. this is pretty crap because i'm going to have to change it for all the languages now so i might change my mind about this later.

deleted some unused views.