Make toggle aspect call synchronous and bump capybaras wait timeout
in features/desktop/aspect_navigation.feature:18 we quickly leave the page
before waiting for the request to finish. By making the HTTP call
synchronous we wait for it to finish before we reflect the change in the DOM,
thus giving Capybara a chance to wait for it. Also let Capybara wait longer.
This should random failures of the aforementioned cuke rarer.