Change markdown-content hr style #5987
The <hr> tag in the markdown-content was #eee,
and thus almost invisible. It has been changed to
$border-gray (currently #ddd) to be consistent
with the other <hr> tags.

Adjust markdown-content hr style further

The hr tag is now centered at 85% width. It also
adds a default vertical margin that is equivalent
to the 0.8em the paragraph tags provide.

Refactor scss for markdown-content hr

Closes #6016