Add DatabaseSizeControl sidetiq job
Cleans initially author signatures from old Like and Participations tables. These signatures take a huge amount of space in the database and their only function is to allow retraction of likes and participations. Allow podmins to define a set of days that these are kept, by default the cleaning is off with a default of 365 days.

Closes #4920

TODO:

* [ ] Do default terms need a if to print out some text regarding cleaning of data if enabled?