Rewrite migration for participation counter
This migrations runs only if the old 20150404193023_participation_counter
migration has not been run.

The new migration is a lot faster, because it runs after the
participation-cleanup migration has been run, so the participation table
is now a lot smaller for people upgrading from 0.5.x to 0.6.x.

It also doesn't remove manually created participations (without likes,
comments or poll participations on the same post).

Fixes #6786