Added button to mobile stream.
Button to stream is now available in the dropdown menu in mobile view.

Updated stream to point to public.

Updated stream to point to public instead of stream

Buttons for stream & public

Buttons for stream and public in the dropdown menu

Added tests for buttons

Added tests for "Stream" and "Public activity" buttons in drawer on mobile.

Changed font-size in drawer

Changed font-size in mobile drawer since more buttons added to list.

Buttons for mobile view streams

Commits for mobile views buttons for streams