Set up Travis to test 4x30 reps of desktop/signs_up.feature:56
This is not a PR. We're only setting up Travis to test a 120 repetitions
of features/desktop/signs_up.feature:56. This will provide us with
sufficient data from CI to determine whether or not our fix is actually
fixing the problem or not.

Would test this locally under normal circumstances but these failures
seem to be reproducable only on Travis.