Upgrade from FactoryGirl 4.9 to FactoryBot 4.11.1
FactoryGirl was renamed FactoryBot, so we need to replace all usage of "FactoryGirl" with "FactoryBot" to upgrade to newer releases of the gem. For that, I just did a find-and-replace throughout the codebase.

Support for static attributions (eg, `gender "robot"`) in factories also got removed in FactoryBot 5.0, so I've converted to using blocks (`gender { "robot" }`) throughout. I tried upgrading straight to to the latest release, but there 3 specs that broke and I'm not yet sure which releases break compatibility or what the fixes are.